chore: 주석 추가, 일부 이름 변경 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSection.tsx 	renamed:    src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.tsx 	renamed:    src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton.type.tsx -> src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.type.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionClothesButton.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionMaskButton.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionMessageBox.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionMessageBox.type.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionMessageBoxLabel.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionSuncreamButton.tsx 	modified:   src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionUmbrellaButton.tsx
