+++ src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.tsx
@@ -3,25 +3,31 @@ import { Pressable, View } from 'react-native';
 
 import classNames from 'classnames';
 
-import type { ChecklistSectionButtonProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton.type';
+import type { ChecklistSectionButtonWrapperProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.type';
 
-export default function ChecklistSectionButton({
+/**
+ * 체크리스트 버튼 래퍼 컴포넌트
+ * @param isSelected 버튼이 선택되었는지 여부
+ * @param children 표시할 아이콘
+ * @jinhok96 25.06.06
+ */
+export default function ChecklistSectionButtonWrapper({
   children,
-  isSelected: selected,
+  isSelected,
   ...props
-}: ChecklistSectionButtonProps) {
+}: ChecklistSectionButtonWrapperProps) {
   const [height, setHeight] = useState(0);
 
   const iconClassName = classNames(
     'aspect-square rounded-xl border',
-    !selected && 'border-background-03 bg-background-03',
-    selected && 'border-morning bg-checklist',
+    !isSelected && 'border-background-03 bg-background-03',
+    isSelected && 'border-morning bg-checklist',
   );
 
   const dashClassName = classNames(
     'absolute left-1/2 border-r border-dashed border-morning transition-[opacity,height]',
-    !selected && 'opacity-0 h-0',
-    selected && 'opacity-100 h-6',
+    !isSelected && 'opacity-0 h-0',
+    isSelected && 'opacity-100 h-6',
   );
 
   return (
