+++ src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.type.tsx
@@ -2,6 +2,6 @@ import type { PressableProps } from 'react-native';
 
 export type ChecklistType = 'umbrella' | 'mask' | 'clothes' | 'suncream';
 
-export type ChecklistSectionButtonProps = Omit<PressableProps, 'className'> & {
+export type ChecklistSectionButtonWrapperProps = Omit<PressableProps, 'className'> & {
   isSelected: boolean;
 };
