+++ src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionClothesButton.tsx
@@ -1,15 +1,15 @@
 import ChecklistLongSleeveIcon from '@components/icon/checklist/ChecklistLongSleeveIcon';
-import ChecklistSectionButton from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton';
+import ChecklistSectionButtonWrapper from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper';
 
-import type { ChecklistSectionButtonProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton.type';
+import type { ChecklistSectionButtonWrapperProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.type';
 
-type ChecklistSectionClothesButtonProps = Omit<ChecklistSectionButtonProps, 'children'>;
+type ChecklistSectionClothesButtonProps = Omit<ChecklistSectionButtonWrapperProps, 'children'>;
 
 export default function ChecklistSectionClothesButton({ ...props }: ChecklistSectionClothesButtonProps) {
   // 복장 추가
   return (
-    <ChecklistSectionButton {...props}>
+    <ChecklistSectionButtonWrapper {...props}>
       <ChecklistLongSleeveIcon />
-    </ChecklistSectionButton>
+    </ChecklistSectionButtonWrapper>
   );
 }
