+++ src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionMessageBox.type.tsx
@@ -1,5 +1,5 @@
 import type { PretendardTextProps } from '@components/fontText/PretendardText.type';
-import type { ChecklistType } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton.type';
+import type { ChecklistType } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.type';
 
 export type ChecklistSectionMessageBoxProps = Omit<PretendardTextProps, 'className' | 'typo'> & {
   selected: ChecklistType | null;
