+++ src/screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionSuncreamButton.tsx
@@ -1,14 +1,14 @@
 import ChecklistSuncreamIcon from '@components/icon/checklist/ChecklistSuncreamIcon';
-import ChecklistSectionButton from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton';
+import ChecklistSectionButtonWrapper from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper';
 
-import type { ChecklistSectionButtonProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButton.type';
+import type { ChecklistSectionButtonWrapperProps } from '@screens/HomeScreen/CurrentForecastScreen/_components/checklistSection/ChecklistSectionButtonWrapper.type';
 
-type ChecklistSectionSuncreamButtonProps = Omit<ChecklistSectionButtonProps, 'children'>;
+type ChecklistSectionSuncreamButtonProps = Omit<ChecklistSectionButtonWrapperProps, 'children'>;
 
 export default function ChecklistSectionSuncreamButton({ ...props }: ChecklistSectionSuncreamButtonProps) {
   return (
-    <ChecklistSectionButton {...props}>
+    <ChecklistSectionButtonWrapper {...props}>
       <ChecklistSuncreamIcon />
-    </ChecklistSectionButton>
+    </ChecklistSectionButtonWrapper>
   );
 }
